Relocate test helpers
