+++ convex/__tests__/earlyAccessCodes.test.ts
@@ -2,8 +2,8 @@ import { convexTest } from "convex-test";
 import { describe, expect, it } from "vitest";
 import { api } from "../_generated/api";
 import schema from "../schema";
-import { createTestAdmin, createTestUser } from "../test-helpers";
 import { modules } from "../test.setup";
+import { createTestAdmin, createTestUser } from "./helpers";
 
 describe("earlyAccessCodes", () => {
   it("creates an early access code", async () => {
