+++ convex/__tests__/userQuests.test.ts
@@ -3,8 +3,8 @@ import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { vi } from "vitest";
 import { api } from "../_generated/api";
 import schema from "../schema";
-import { createTestUser } from "../test-helpers";
 import { modules } from "../test.setup";
+import { createTestUser } from "./helpers";
 
 const UPDATE_TIMESTAMP = Date.now();
