+++ convex/__tests__/userSettings.test.ts
@@ -2,8 +2,8 @@ import { convexTest } from "convex-test";
 import { describe, expect, it } from "vitest";
 import { api } from "../_generated/api";
 import schema from "../schema";
-import { createTestUser } from "../test-helpers";
 import { modules } from "../test.setup";
+import { createTestUser } from "./helpers";
 
 describe("userSettings", () => {
   describe("getCurrentUserSettings", () => {
